app: clean up services and separate api from src into apps.
