+++ src/memoryfm/io/lastfmstats.py
@@ -5,18 +5,14 @@ from zoneinfo import ZoneInfo
 
 from memoryfm.errors import SchemaError, InvalidDataError
 from memoryfm.io._loaders import load_json
-from memoryfm.util.datetime_util import validate_tz
-from memoryfm.db_services.basic_fetch import (
-    create_user,
-    get_userid_from_username,
-    insert_scrobbles,
-)
+import memoryfm.services.user_service as userv
+import memoryfm.services.scrobble_service as scserv
 
 if TYPE_CHECKING:
     from pathlib import Path
 
 
-def validate_lastfmstats(file: str | Path, tz: str | None = None):
+def validate_lastfmstats(file: str | Path):
     data = load_json(file)
     if not data:
         raise InvalidDataError("JSON data is empty")
@@ -26,23 +22,26 @@ def validate_lastfmstats(file: str | Path, tz: str | None = None):
                 raise SchemaError(f"Key missing in JSON file: '{key}'", key)
         username = data.get("username")
         scrobbles_data = data.get("scrobbles")
-        tz = validate_tz(tz)
-        return username, scrobbles_data, tz
+        return username, scrobbles_data
 
 
 def from_lastfmstats(file: str | Path, tz: str | None = None, overwrite: bool = False):
-    username, scrobbles_data, tz = validate_lastfmstats(file, tz)
+    username, scrobbles_data = validate_lastfmstats(file)
     if username and scrobbles_data:
-        scrobbles = [
-            {
-                "timestamp": datetime.fromtimestamp(s["date"] / 1000, tz=ZoneInfo(tz)),
-                "artist": s["artist"],
-                "album": s["album"],
-                "track": s["track"],
-            }
-            for s in scrobbles_data
-        ]
-        create_user(username, tz, overwrite)
-        user_id = get_userid_from_username(username)
-        if user_id:
-            insert_scrobbles(user_id, scrobbles)
+        userv.create_user(username, tz, overwrite)
+        context = userv.get_user_context(username)
+        user_id = context.get("user_id") if context else None
+        tz = context.get("tz") if context else tz
+        if user_id and tz:
+            scrobbles = [
+                {
+                    "timestamp": datetime.fromtimestamp(
+                        s["date"] / 1000, tz=ZoneInfo(tz)
+                    ),
+                    "artist": s["artist"],
+                    "album": s["album"],
+                    "track": s["track"],
+                }
+                for s in scrobbles_data
+            ]
+            scserv.insert_scrobbles(user_id, scrobbles)
